null type was being fetched by the VersionNeutralBinder
Used by the BinaryFormatter. It probably broke because our dlls aren't signed, so the regex was failing.

- Update regex to handle 'null' PublicKeyToken in type names

The regular expression in the `VersionNeutralBinder` class has been
modified to allow the `PublicKeyToken` to be either a sequence of
hexadecimal digits or the string "null". This change ensures that
type names with a `PublicKeyToken` of "null" are correctly processed
by the `BindToType` method.

- Add fallback to BindToType for resolving types with assembly

Updated BindToType method in VersionNeutralBinder to handle
cases where Type.GetType(typeName) returns null. If the type is not found, it now attempts to resolve the type by appending the assembly name to the type name.