Fix bug where files where being marked as dirty because of different line endings.
The reason is that when opening the `ComicBookDialog` some multiline fields are modified so that they appear correctly in the dialog. This change triggers to mark the file as requiring an update.

This change adds a step to compare normallized strings taking into consideration line endings before trigerring the `BookChanged` event.

close #161
---

Enhance multiline text handling and equality checks

- Introduced `CheckMultilineEquality` method in `ComicInfo.cs` for improved property comparison with string normalization.
- Updated `ComicBookDialog.cs` to use `StringUtility.MakeEditBoxMultiline` for `Summary`, `Notes`, and `Review` properties. Ensuring no duplicate are being added to the `AutoCompleteSource` of the textbox because of line endings differences.
- Added `NormalizeLineEndings` method in `StringUtility.cs` to standardize line endings for accurate text comparisons.