Prevent the top book of a stack made using Virtual tags to change seamlessly at random.
Stacking using Virtual Tags will now additionally sort the stack by series and should now ignore case & articles.

When stacking the field used to stack is the same than the field used to sort the stack.  Since the value is the same for the whole stack it doesn't permit to properly sort the stack. So now for the sorting for virtual tags will additionally sort using the series comparer (i.e. Virtual Tag->Series->Format->Volume->Number).

This only affects virtual tags and will not change when stacking via other properties. For example stacking by Publisher or Series Group will not change the random cover changing. This requires updating each fields way of sorting.

The reason the top of the stack keeps changing is that when the items are added to the stack, they are done in parallel (unless disabled) so it is possible that some items are added in different orders. The program does sort them afterward, but like with the same stack value that doesn't change. To fix this completely there would be a need for an additional sort to determine the top of the stack. For example sorting a Publisher or a Virtual Tags stack via Published Date is currently not possible.