Stack order is now determined by the sort from inside the stack
Items in a stack are usually sorted by the field used to stack.
For types like text (ex: Series Group) what happens is that all the value are the same.
Other like series stacking/sorting are different because they compare multiple fields, like Format, Volume, Series & Number ensuring they are always in order.
When there is no specific order the item will be in the order they were added to the stack. But since this operation is done in parallel the order they are added isn't consistent.
Ending up in a situation where the top of the stack changes constantly.

So in cases like these were the values are always the same, these stacks will now be sorted by the setting inside the stack.
This will let you easily choose how you want the top of the stack be set. It does not take into consideration grouping inside the stack, only the sort setting.

There is a setting in parameters that makes it so every stack in a layout is always the same.
With that option enabled all stack in a list can have different config, but there is a limit of 1024 different config per list.
If for any reason a config doesn't exists the program will default to the series comparer.

See issue #68