Added an "escape" Virtual Tag function
This is useful when you need to reference some field when using an expression parameter. Since there is no way for an expression to be set as a verbatim literal string (prefixing a string w/ @). this is required to escape strings that contains character that are usually escaped in C#.

Ex: It would be impossible to get the length of a file path by using `[$expr<"{FilePath}".Length>]` because the `\` in a path would not be correctly escaped when evaluating the C# string.