Inverted default value for OptimizedListScrolling
This would basically always be enabled anyway on modern systems. But if you would change some hardware settings in ComicRack.ini it would bypass the Auto-Tune (or if disabled). Bypassing Auto-Tune would mean that the OptimizedListScrolling was never disabled, giving severe pop-in of covers while scrolling. Upping the texture memory would result in worst behavior.

This way it is always disabled and only enables on slower machine instead, rather than the other way around. Like the rest of the other related settings that affect Auto-Tune. They usually change a setting for lower performance machine and default to the highest quality, this changes align this setting with the other settings.