Prevent smart list using Virtual Tags from using from using a cache.
- Added a new function to ReverseIndex to use a more general Where.
- Used new function to always get lists that used VirtualTag as properties and invalidate their cache.
- Made all virtual tag Matchers DisableOptimizedUpdate true so that they always reset the cache.

NOTE: The best solution would be for the virtual tag to have their own reverse index that lists what properties they reference. This would require parsing the virtual tags and storing the properties used before the caption is ran by GetFullTitle. Instead we are now resetting the cache for every virtual tag smart list when we change any values.