Backup manager (#218)
Integrated a Backup Manager that backups on Startup & Exit. This is to replace the Backup Manager plugin. This isn't a replacement for the current Backup Database feature. They co-exists at the same time. It is also not possible to restore from that backup using the UI.

Difference between the plugin are:

    - It will automatically delete old backups when over the specified quantity to keep (they are sent to the recycle bin).
    - The Include all Alternate Configs means that instead of only backing up only the current config, it will save all configs present in the program data folder regardless of which is loaded. This is the equivalent to the old "Full" backup from the plugin.
    - You can pick and choose which elements you want in your backup.
    - When backing up on exit it does so after the database have been saved. This is after plugins are triggered, so you will have the correct version of your database. The plugin would trigger before so you wouldn't have the very latest database backup.
    - It also backups files from the `%ProgramData%` directory (but never the ones from the startup directory).
    - The backup is done in the background so it doesn't block the UI on startup like the plugin does (unless you used the buggy v1.3)

Go in Preferences -> Advanced -> Backup Manager. Enable either the on Startup or Exit option and some elements to save.

Note: For user using SQL, the current Backup Database feature saves the contents of the books from the db into the backup. This DOES NOT, it just backups your current XML file.

Setting the # of Backups to Keep to 0 will disable deletion of old backups. Do not rename the files or have multiple backups from different alternate configuration in the same folder. Finding old files is done by checking the filename and doesn't check subfolders.